Refactor mzn_models and add impossible differential model based on infeasibility
